# internal/docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.loginRequest:
    properties:
      login:
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
  auth.refreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  auth.registerRequest:
    properties:
      login:
        type: string
      password:
        type: string
      role:
        type: string
    required:
    - login
    - password
    type: object
  bid.PlaceBidRequest:
    properties:
      amount:
        type: number
    required:
    - amount
    type: object
  lots.lotsRequest:
    properties:
      category:
        type: string
      description:
        type: string
      endsAt:
        type: string
      photo:
        type: string
      startPrice:
        type: number
      status:
        type: string
      title:
        type: string
    required:
    - category
    - description
    - endsAt
    - photo
    - startPrice
    - status
    - title
    type: object
  lots.moderateRequest:
    properties:
      approve:
        type: boolean
      reason:
        type: string
    type: object
  lots.pagResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/model.Lots'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
  lots.updateLotRequest:
    properties:
      ID:
        type: integer
      category:
        type: string
      description:
        type: string
      endsAt:
        type: string
      photo:
        type: string
      startPrice:
        type: number
      title:
        type: string
    required:
    - category
    - description
    - endsAt
    - photo
    - startPrice
    - title
    type: object
  model.Bid:
    properties:
      amount:
        type: number
      created_at:
        type: string
      lot_id:
        type: integer
      lot_title:
        type: string
    type: object
  model.Lots:
    properties:
      category:
        type: string
      createdAt:
        type: string
      currentPrice:
        format: float64
        type: number
      description:
        type: string
      endAt:
        type: string
      id:
        format: int64
        type: integer
      moderationStatus:
        type: string
      photo:
        type: string
      rejectionReason:
        type: string
      sellerID:
        format: int64
        type: integer
      startAt:
        type: string
      startPrice:
        format: float64
        type: number
      status:
        type: string
      title:
        type: string
      winnerID:
        format: int64
        type: integer
    type: object
  model.Win:
    properties:
      closed_at:
        type: string
      lot_id:
        type: integer
      lot_title:
        type: string
      winning_bid:
        type: number
    type: object
host: localhost:9999
info:
  contact: {}
  title: AuctionHouse API
  version: "1.0"
paths:
  /admin/close-expired:
    post:
      description: Ручной запуск закрытия лотов с истёкшим дедлайном
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
      security:
      - BearerAuth: []
      summary: Закрыть просроченные лоты
      tags:
      - admin
  /admin/lots:
    get:
      description: Возвращает все лоты с фильтрацией по статусу/продавцу/дате, поиск,
        пагинация. Доступна только админу
      parameters:
      - description: Поиск по названию лота
        in: query
        name: search
        type: string
      - description: Фильтр по статусу (draft, live, closed, cancelled)
        in: query
        name: status
        type: string
      - description: Фильтр по ID продавца
        in: query
        name: seller_id
        type: integer
      - description: Номер страницы (по умолчанию 1)
        in: query
        name: page
        type: integer
      - description: Размер страницы (по умолчанию 50, максимум 100)
        in: query
        name: page_size
        type: integer
      - description: 'Поле для фильтрации по дате: created_at или ends_at(default
          created_at)'
        in: query
        name: date_field
        type: string
      - description: Начало периода (формат YYYY-MM-DD)
        in: query
        name: date_from
        type: string
      - description: Конец периода (формат YYYY-MM-DD)
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lots.pagResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получает лот по title, status, seller_id, date
      tags:
      - admin
  /admin/lots/{id}/moderate:
    put:
      consumes:
      - application/json
      description: Одобрить или отклонить лот перед публикацией. При отклонении требуется
        причина.
      parameters:
      - description: ID лота
        in: path
        name: id
        required: true
        type: integer
      - description: Решение модерации
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/lots.moderateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Модерация лота
      tags:
      - admin
  /auth/login:
    post:
      consumes:
      - application/json
      description: Проверяет логин и пароль, возвращает токен
      parameters:
      - description: Логин и пароль
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/auth.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Авторизация пользователя
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Отзывает refresh-токен
      parameters:
      - description: Refresh токен
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/auth.refreshRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Выход пользователя
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Ротирует refresh-токен и выдаёт новую пару токенов
      parameters:
      - description: Refresh токен
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/auth.refreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
      summary: Обновление токенов
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Создаёт пользователя в базу
      parameters:
      - description: Регистрация
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/auth.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Регистрация пользователя
      tags:
      - auth
  /bids:
    get:
      description: Возвращает все ставки авторизованного участника
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Bid'
            type: array
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Получить свои ставки
      tags:
      - bids
  /lots:
    get:
      description: Метод возрвщает список активных лотов
      parameters:
      - description: Поиск
        in: query
        name: search
        type: string
      - description: Категория
        in: query
        name: category
        type: string
      - description: Минимальная цена
        in: query
        name: minPrice
        type: number
      - description: Максимальная цена
        in: query
        name: maxPrice
        type: number
      - description: Номер страницы
        in: query
        name: page
        type: integer
      - description: Количество элементов
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      security:
      - BearerAuth: []
      summary: Получение список лотов
      tags:
      - lots
  /lots/:id:
    delete:
      description: Получает лот по ID и удаляет его
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
      security:
      - BearerAuth: []
      summary: Удаление лот по ID
      tags:
      - lots
    get:
      description: Returns a lot by its ID
      parameters:
      - description: Лот ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
      security:
      - BearerAuth: []
      summary: Получение лота по ID
      tags:
      - lots
    put:
      consumes:
      - application/json
      description: Получает лот по ID и меняет его значение
      parameters:
      - description: Изменит лот
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/lots.updateLotRequest'
      - description: ID лота
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
      security:
      - BearerAuth: []
      summary: Получение лот по ID
      tags:
      - lots
  /lots/:id/bid:
    post:
      consumes:
      - application/json
      description: Размещает ставку на активный лот
      parameters:
      - description: ID лота
        in: path
        name: id
        required: true
        type: integer
      - description: Сумма ставки
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/bid.PlaceBidRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "409":
          description: Conflict
      security:
      - BearerAuth: []
      summary: Сделать ставку
      tags:
      - bids
  /lots/:id/photo:
    get:
      description: По ID получает фото из БД
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
      security:
      - BearerAuth: []
      summary: Получение фото лота
      tags:
      - lots
    put:
      description: Проверяет размер фото и обновляет путь фото в БД
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
      security:
      - BearerAuth: []
      summary: Обновление фото
      tags:
      - lots
  /lots/{id}/status:
    put:
      description: Меняет статус лота draft→live→closed->cancelled
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
      security:
      - BearerAuth: []
      summary: Смена статуса
      tags:
      - lots
  /lots/new:
    post:
      description: Создает лоты в базу
      parameters:
      - description: Добавить лот
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/lots.lotsRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
      security:
      - BearerAuth: []
      summary: Создать лоты
      tags:
      - lots
  /me:
    get:
      description: Возвращает данные авторизованного пользователя.
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Информация о текущем пользователе
      tags:
      - user
  /wins:
    get:
      description: Возвращает закрытые лоты, выигранные авторизованным участником
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Win'
            type: array
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Получить выигранные лоты
      tags:
      - wins
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"